Deduplicate implementation of misaligned loads and stores
Two policies uses the misaligned loads and stores emulation and we have code duplication at the moment. This commit extracts the logic in a third file and deduplicates the code.
