+++ src/policy/offload.rs
@@ -3,13 +3,11 @@
 
 use miralis_core::sbi_codes;
 
-use crate::arch::{
-    get_raw_faulting_instr, parse_mpp_return_mode, Arch, Architecture, Csr, MCause, Register,
-};
-use crate::decoder::Instr;
+use crate::arch::{get_raw_faulting_instr, Arch, Architecture, Csr, MCause, Register};
 use crate::host::MiralisContext;
 use crate::platform::{Plat, Platform};
 use crate::policy::{PolicyHookResult, PolicyModule};
+use crate::rw_emulation::{emulate_misaligned_read, emulate_misaligned_write};
 use crate::virt::traits::{RegisterContextGetter, RegisterContextSetter};
 use crate::virt::VirtContext;
 
@@ -32,8 +30,8 @@ impl PolicyModule for OffloadPolicy {
         let trap_info = ctx.trap_info.clone();
 
         match trap_info.get_cause() {
-            MCause::LoadAddrMisaligned => self.emulate_misaligned_read(ctx, mctx),
-            MCause::StoreAddrMisaligned => self.emulate_misaligned_write(ctx, mctx),
+            MCause::LoadAddrMisaligned => emulate_misaligned_read(ctx, mctx),
+            MCause::StoreAddrMisaligned => emulate_misaligned_write(ctx, mctx),
             MCause::EcallFromSMode => {
                 let timer_eid: bool = ctx.get(Register::X17) == sbi_codes::SBI_TIMER_EID;
                 let timer_fid: bool = ctx.get(Register::X16) == sbi_codes::SBI_TIMER_FID;
@@ -78,148 +76,4 @@ impl PolicyModule for OffloadPolicy {
     }
 
     const NUMBER_PMPS: usize = 0;
-}
-
-impl OffloadPolicy {
-    fn emulate_misaligned_read(
-        &mut self,
-        ctx: &mut VirtContext,
-        mctx: &mut MiralisContext,
-    ) -> PolicyHookResult {
-        let instr_ptr = ctx.trap_info.mepc as *const u8;
-
-        let mut instr: [u8; 8] = [0, 0, 0, 0, 0, 0, 0, 0];
-        unsafe {
-            self.copy_from_previous_mode(instr_ptr, &mut instr);
-        }
-
-        match mctx.decode_load(u64::from_le_bytes(instr) as usize) {
-            Instr::Load {
-                rd, rs1, imm, len, ..
-            } => {
-                // Build the value
-                let start_addr: *const u8 =
-                    ((ctx.regs[rs1 as usize] as isize + imm) as usize) as *const u8;
-
-                match len.to_bytes() {
-                    8 => {
-                        let mut value_to_read: [u8; 8] = [0, 0, 0, 0, 0, 0, 0, 0];
-                        unsafe {
-                            self.copy_from_previous_mode(start_addr, &mut value_to_read);
-                        }
-
-                        // Return the value
-                        ctx.regs[rd as usize] = u64::from_le_bytes(value_to_read) as usize;
-                    }
-                    4 => {
-                        let mut value_to_read: [u8; 4] = [0, 0, 0, 0];
-                        unsafe {
-                            self.copy_from_previous_mode(start_addr, &mut value_to_read);
-                        }
-
-                        // Return the value
-                        ctx.regs[rd as usize] = u32::from_le_bytes(value_to_read) as usize;
-                    }
-                    2 => {
-                        let mut value_to_read: [u8; 2] = [0, 0];
-                        unsafe {
-                            self.copy_from_previous_mode(start_addr, &mut value_to_read);
-                        }
-
-                        // Return the value
-                        ctx.regs[rd as usize] = u16::from_le_bytes(value_to_read) as usize;
-                    }
-                    _ => {
-                        todo!("Implement support for other than 2,4,8 bytes misalinged accesses, current size: {}", len.to_bytes())
-                    }
-                }
-            }
-            _ => {
-                panic!("Must be a load instruction here")
-            }
-        }
-
-        ctx.pc += 4;
-        PolicyHookResult::Overwrite
-    }
-
-    fn emulate_misaligned_write(
-        &mut self,
-        ctx: &mut VirtContext,
-        mctx: &mut MiralisContext,
-    ) -> PolicyHookResult {
-        let instr_ptr = ctx.trap_info.mepc as *const u8;
-
-        let mut instr: [u8; 8] = [0, 0, 0, 0, 0, 0, 0, 0];
-        unsafe {
-            self.copy_from_previous_mode(instr_ptr, &mut instr);
-        }
-
-        match mctx.decode_store(u64::from_le_bytes(instr) as usize) {
-            Instr::Store {
-                rs2, rs1, imm, len, ..
-            } => {
-                assert!(
-                    len.to_bytes() == 8 || len.to_bytes() == 4 || len.to_bytes() == 2,
-                    "Implement support for other than 2,4,8 bytes misalinged accesses"
-                );
-
-                if rs1 as usize != ((rs1 as usize) as isize) as usize {
-                    panic!("Conversion is weird")
-                }
-
-                // Build the value
-                let start_addr: *mut u8 =
-                    ((ctx.regs[rs1 as usize] as isize + imm) as usize) as *mut u8;
-
-                match len.to_bytes() {
-                    8 => {
-                        let val = ctx.regs[rs2 as usize] as u64;
-                        let mut value_to_store: [u8; 8] = val.to_le_bytes();
-
-                        unsafe {
-                            self.copy_from_previous_mode_store(&mut value_to_store, start_addr);
-                        }
-                    }
-                    4 => {
-                        let val = ctx.regs[rs2 as usize] as u32;
-                        let mut value_to_store: [u8; 4] = val.to_le_bytes();
-
-                        unsafe {
-                            self.copy_from_previous_mode_store(&mut value_to_store, start_addr);
-                        }
-                    }
-                    2 => {
-                        let val = ctx.regs[rs2 as usize] as u16;
-                        let mut value_to_store: [u8; 2] = val.to_le_bytes();
-
-                        unsafe {
-                            self.copy_from_previous_mode_store(&mut value_to_store, start_addr);
-                        }
-                    }
-                    _ => {
-                        todo!("Implement support for other than 2,4,8 bytes misalinged accesses, current size: {}", len.to_bytes())
-                    }
-                }
-            }
-            _ => {
-                panic!("Must be a load instruction here")
-            }
-        }
-
-        ctx.pc += 4;
-        PolicyHookResult::Overwrite
-    }
-
-    unsafe fn copy_from_previous_mode(&mut self, src: *const u8, dest: &mut [u8]) {
-        // Copy the arguments from the S-mode virtual memory to the M-mode physical memory
-        let mode = parse_mpp_return_mode(Arch::read_csr(Csr::Mstatus));
-        unsafe { Arch::read_bytes_from_mode(src, dest, mode).unwrap() }
-    }
-
-    unsafe fn copy_from_previous_mode_store(&mut self, src: &mut [u8], dest: *mut u8) {
-        // Copy the arguments from the S-mode virtual memory to the M-mode physical memory
-        let mode = parse_mpp_return_mode(Arch::read_csr(Csr::Mstatus));
-        unsafe { Arch::store_bytes_from_mode(src, dest, mode).unwrap() }
-    }
 }
